+++ CHANGELOG.md
@@ -197,7 +197,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- block Lanczos must return hermitian matrices ([#42](https://github.com/frankebel/RAS_DMFT.jl/pull/42)) ([d0b066a](https://github.com/frankebel/RAS_DMFT.jl/commit/d0b066aba90a8308ea0f9adeece25165e52acaba))
+- block Lanczos must return Hermitian matrices ([#42](https://github.com/frankebel/RAS_DMFT.jl/pull/42)) ([d0b066a](https://github.com/frankebel/RAS_DMFT.jl/commit/d0b066aba90a8308ea0f9adeece25165e52acaba))
 - correlator FR in improved self-energy ([#51](https://github.com/frankebel/RAS_DMFT.jl/pull/51)) ([f374b44](https://github.com/frankebel/RAS_DMFT.jl/commit/f374b444381c9c2a612561cf0d95c32a1733dff3))
 
 ## [0.5.1] - 2025-04-02